+++ test/PHPUnit/connecteur/generateur-seda/SedaGeneriqueTest.php
@@ -1,7 +1,5 @@
 <?php
 
-use Twig\Error\RuntimeError;
-
 require_once __DIR__ . "/../../../../connecteur/generateur-seda/lib/FluxDataTestSedaGenerique.class.php";
 
 class SedaGeneriqueTest extends PastellTestCase
@@ -195,7 +193,33 @@ class SedaGeneriqueTest extends PastellTestCase
         $id_ce = $this->createSedaGeneriqueConnector();
         $sedaGeneriqueConnector = $this->getConnecteurFactory()->getConnecteurById($id_ce);
         $this->expectException(UnrecoverableException::class);
-        $this->expectExceptionMessage("SedaGenerator did not return a 200 response code (503 instead).");
+        $this->expectExceptionMessage("SedaGenerator did not return a 200 response. Code HTTP: 503.");
         $sedaGeneriqueConnector->getBordereauNG(new FluxDataTestSedaGenerique());
+    }
+
+    /**
+     * @throws UnrecoverableException
+     */
+    public function testWhenConnectionIsOk()
+    {
+        $this->mockCurl(["http://seda-generator:8080/ping" => "OK"]);
+        $id_ce = $this->createSedaGeneriqueConnector();
+        /* @var SedaGenerique $sedaGeneriqueConnector */
+        $sedaGeneriqueConnector = $this->getConnecteurFactory()->getConnecteurById($id_ce);
+        $this->assertTrue($sedaGeneriqueConnector->testConnexion());
+    }
+
+    /**
+     * @throws UnrecoverableException
+     */
+    public function testWhenConnectionIsNotOk()
+    {
+        $this->mockCurl(["http://seda-generator:8080/ping" => "OK"], 404);
+        $id_ce = $this->createSedaGeneriqueConnector();
+        /* @var SedaGenerique $sedaGeneriqueConnector */
+        $sedaGeneriqueConnector = $this->getConnecteurFactory()->getConnecteurById($id_ce);
+        $this->expectException(UnrecoverableException::class);
+        $this->expectExceptionMessage("SedaGenerator did not return a 200 response. Code HTTP: 404.");
+        $sedaGeneriqueConnector->testConnexion();
     }
 }